+++ lightly_studio/src/lightly_studio/api/routes/api/image/routes.py
@@ -1,0 +1,198 @@
+"""API routes for managing image samples."""
+
+from __future__ import annotations
+
+from typing import Annotated
+from uuid import UUID
+
+from fastapi import APIRouter, Depends, HTTPException, Path, Query
+from pydantic import BaseModel, Field
+
+from lightly_studio.api.routes.api.collection import get_and_validate_collection_id
+from lightly_studio.api.routes.api.image.count_by_sample_tags import (
+    count_by_sample_tags_router,
+)
+from lightly_studio.api.routes.api.status import HTTP_STATUS_NOT_FOUND
+from lightly_studio.api.routes.api.validators import Paginated
+from lightly_studio.database.db_manager import SessionDep
+from lightly_studio.models.annotation.annotation_base import AnnotationType
+from lightly_studio.models.collection import CollectionTable
+from lightly_studio.models.image import ImageView, ImageViewsWithCount
+from lightly_studio.models.sort import SortExpr, sort_expr_to_order_by
+from lightly_studio.resolvers import (
+    image_resolver,
+)
+from lightly_studio.resolvers.image_filter import (
+    ImageFilter,
+)
+from lightly_studio.resolvers.image_resolver.annotation_count_types import (
+    AnnotationCountMode,
+)
+
+image_router = APIRouter(tags=["image"])
+image_router.include_router(
+    count_by_sample_tags_router,
+    prefix="/collections/{collection_id}/images",
+)
+
+
+class ReadImagesRequest(BaseModel):
+    """Request body for reading samples with text embedding."""
+
+    filters: ImageFilter | None = Field(None, description="Filter parameters for samples")
+    text_embedding: list[float] | None = Field(None, description="Text embedding to search for")
+    sample_ids: list[UUID] | None = Field(None, description="The list of requested sample IDs")
+    pagination: Paginated | None = Field(
+        None, description="Pagination parameters for offset and limit"
+    )
+    sort_by: list[SortExpr] | None = Field(None, description="Sort expressions for ordering")
+
+
+@image_router.post("/collections/{collection_id}/images/list")
+def read_images(
+    session: SessionDep,
+    collection_id: Annotated[UUID, Path(title="collection Id")],
+    body: ReadImagesRequest,
+) -> ImageViewsWithCount:
+    """Retrieve a list of samples from the database with optional filtering.
+
+    Args:
+        session: The database session.
+        collection_id: The ID of the collection to filter samples by.
+        body: Optional request body containing text embedding.
+
+    Returns:
+        A list of filtered samples.
+    """
+    order_by = [sort_expr_to_order_by(expr) for expr in body.sort_by] if body.sort_by else None
+    result = image_resolver.get_all_by_collection_id(
+        session=session,
+        collection_id=collection_id,
+        pagination=body.pagination,
+        filters=body.filters,
+        text_embedding=body.text_embedding,
+        sample_ids=body.sample_ids,
+        order_by=order_by,
+    )
+    # TODO(Michal, 10/2025): Add SampleView to ImageView and then use a response model
+    # instead of manual conversion.
+    scores: list[float | None] = (
+        list(result.similarity_scores) if result.similarity_scores else [None] * len(result.samples)
+    )
+    order_values: list[float | None] = (
+        list(result.order_values) if result.order_values else [None] * len(result.samples)
+    )
+    return ImageViewsWithCount(
+        samples=[
+            ImageView.from_image_table(image=image, similarity_score=score, order_value=order_value)
+            for image, score, order_value in zip(result.samples, scores, order_values)
+        ],
+        total_count=result.total_count,
+        next_cursor=result.next_cursor,
+    )
+
+
+@image_router.get("/collections/{collection_id}/images/dimensions")
+def get_image_dimensions(
+    session: SessionDep,
+    collection: Annotated[
+        CollectionTable,
+        Path(title="collection Id"),
+        Depends(get_and_validate_collection_id),
+    ],
+    annotation_label_ids: Annotated[list[UUID] | None, Query()] = None,
+) -> dict[str, int]:
+    """Get min and max dimensions of samples in a collection."""
+    return image_resolver.get_dimension_bounds(
+        session=session,
+        collection_id=collection.collection_id,
+        annotation_label_ids=annotation_label_ids,
+    )
+
+
+@image_router.get("/images/{sample_id}")
+def read_image(
+    session: SessionDep,
+    sample_id: Annotated[UUID, Path(title="Sample Id")],
+) -> ImageView:
+    """Retrieve a single sample from the database."""
+    image = image_resolver.get_by_id(session=session, sample_id=sample_id)
+    if not image:
+        raise HTTPException(status_code=HTTP_STATUS_NOT_FOUND, detail="Sample not found")
+
+    return ImageView.from_image_table(image=image)
+
+
+class SampleAdjacentsParams(BaseModel):
+    """Parameters for getting adjacent samples."""
+
+    filters: ImageFilter | None = None
+    text_embedding: list[float] | None = None
+
+
+class ReadImageSampleIdsRequest(BaseModel):
+    """Request body for reading matching image sample ids."""
+
+    filters: ImageFilter | None = Field(None, description="Filter parameters for images")
+
+
+class ReadCountImageAnnotationsRequest(BaseModel):
+    """Request body for reading image annotation counts."""
+
+    filter: ImageFilter | None = None
+    annotation_type: AnnotationType | None = Field(
+        None,
+        description="Restrict counts to a single annotation type (e.g. classification).",
+    )
+    count_mode: AnnotationCountMode = Field(
+        AnnotationCountMode.OBJECTS,
+        description="Whether to count annotation objects or distinct annotated samples.",
+    )
+
+
+@image_router.post("/collections/{collection_id}/images/sample_ids", response_model=list[UUID])
+def get_image_sample_ids(
+    session: SessionDep,
+    collection_id: Annotated[UUID, Path(title="collection Id")],
+    body: ReadImageSampleIdsRequest,
+) -> list[UUID]:
+    """Retrieve all sample ids of images matching the given filters."""
+    return list(
+        image_resolver.get_sample_ids(
+            session=session,
+            collection_id=collection_id,
+            filters=body.filters,
+        )
+    )
+
+
+@image_router.post("/collections/{collection_id}/images/annotations/count")
+def count_image_annotations_by_collection(
+    collection: Annotated[
+        CollectionTable,
+        Path(title="collection Id"),
+        Depends(get_and_validate_collection_id),
+    ],
+    session: SessionDep,
+    body: ReadCountImageAnnotationsRequest | None = None,
+) -> list[dict[str, str | int]]:
+    """Get image annotation counts for a specific collection."""
+    image_filter = body.filter if body and body.filter else None
+    annotation_type = body.annotation_type if body else None
+    count_mode = body.count_mode if body else AnnotationCountMode.OBJECTS
+    counts = image_resolver.count_image_annotations_by_collection(
+        session=session,
+        collection_id=collection.collection_id,
+        image_filter=image_filter,
+        annotation_type=annotation_type,
+        count_mode=count_mode,
+    )
+
+    return [
+        {
+            "label_name": label_name,
+            "current_count": current_count,
+            "total_count": total_count,
+        }
+        for label_name, current_count, total_count in counts
+    ]
