+++ lightly_studio/tests/resolvers/image_resolver/test_count_image_annotations_by_collection.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-from lightly_studio.resolvers.image_resolver.annotation_count_types import AnnotationCountMode
 from sqlmodel import Session
 
 from lightly_studio.resolvers import image_resolver
 from lightly_studio.resolvers.annotations.annotations_filter import AnnotationsFilter
 from lightly_studio.resolvers.image_filter import ImageFilter
+from lightly_studio.resolvers.image_resolver.annotation_count_types import AnnotationCountMode
 from lightly_studio.resolvers.sample_resolver.sample_filter import SampleFilter
 from tests.helpers_resolvers import (
     AnnotationDetails,
